Modifications to bgvideo html
Removes the <source> child of the <video> element in favor of a src attribute on the <video> tag itself. This is for mobile compatibility. The old version does not work with the latest versions of iOS, even if only the mp4 video type is used. Also changes how the muted, autoplay, and loop attributes are rendered (full boolean), and adds the playsinline attribute.